added unsinged long conversion to avoid integer overflow for starting delay calculation

+++ pitch_machine_feeder_v2/pitch_machine_feeder_v2.ino
@@ -206,7 +206,7 @@ void pitching(){
       if(false == active){
         return;
       }
-      delay(STARTING_DELAY/100);
+      delay(((unsigned long) STARTING_DELAY)/100);
     }
   }
